perf(l1): rebuild_bloom - insert instead of merge (#5266)
Merging layer filters is very costly due to the cost of iterating the
elements of a filter. Inserting has proven to be cheaper for the size of
the filter we're using and the amount of elements we are holding. Some
measurements:

- Early hoodi imports went from 8 blocks per second to about 80 (10x).
- Synced hoodi rebuild time went from ~166ms to ~27ms, (about a 6x). 

The PR also simplifies the rebuild code and avoids using rayon entirely.

---------

Co-authored-by: Javier Rodríguez Chatruc <49622509+jrchatruc@users.noreply.github.com>
Co-authored-by: Javier Chatruc <jrchatruc@gmail.com>