revert(l1, l2): revert revert of build empty block (#5337)
**Description**

This commit reverts the revert of #5205 done in #5257. 
As of #5310, the dev mode block producer now waits for the configured
amount of time after initiating the payload build process and before
retrieving the built payload, rather than after payload validation. This
solves the issue we were having after merging #5205 where empty payloads
were retrieved because the `engine_getPayload` call was immediately done
after the `engine_forkChoiceUpdate`.